Add LegacyCloseables class to replace Closeables.closeQuietly(Closeable)
The Closeables.closeQuietly(Closeable) method is no longer
present in v18 but it is still used in Android code and many of
the usages cannot be replaced with try-with-resources as the
code has to work on Android API versions < 19 which is the first
version at which try-with-resources was supported.

This is part of the work to upgrade Guava to version 18.

BUG: 19672715
Change-Id: Ie4c592e3cdea3554c21ad48a901e37268a939631